Pushed remaining changes
(cherry picked from commit 4f51d99efd4d4f0404a1178815f90e2ebe1b6841)
